First draft
This is what I have done as of yet. It's pretty rough code and not necessarily easy to read through but its a start. My next goal with it is to make sure it has the functionality, design and names etc. based off the use cases before the iteration is due.